refactor(binding-http/server): introduce routes for each TD endpoint
**Note**: that we are now use find-my-way for routing, which is a bit more
complex than the previous approach. However, it is also more flexible
and in the future we can easily add more routes or let user configure
them. Another caveat is the trailing slashes in the routes now map to
a total different path. For example, `/things` and `/things/` are now
different routes. In the future, we might review this behavior.
